add sqlite and a test command (lose WASI support, see note)
With this commit, we lose WASI. This is ok here, because it's not like
we're going to be saving our data on sqlite inside a Cloudflare worker.
The problem actually doesn't seem to be compilation, but runtime.
wasmtime complains that too many locals are defined when sqlite is
compiled into this